# src/main/java/com/betting/dto/request/VoteRequest.java
package com.betting.dto.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class VoteRequest {
    private boolean vote; // true - за, false - против
} 